Fixing compound rule logic
The case of a literal Nil rule as the only member inside an AND rule
inside a NOT rule was not working.  An equals literal must be
satisfiable when the triggering key equals the rule's key and no
other restrictions, because the rule may be inside a compound rule
inside a NOT rule.